storage: update SD role in domain state callback
In commit I2ecf801 (sp: _refreshDomainLinks must not change the
metadata),updating the storage domain role from master to regular, was
moved to startSpm method. This update is needed in case that the master
domain changed. Updating the storage domain role is a metadata
operation, therefore it should be done only by the SPM.

After commit I2ecf801, starting the SPM will fail if some domain is not
accessible.

Some failures are handled in commit Ia4f3436 (sp: fix spm start when
failing to produce domain), that handled only  StorageDomainDoesNotExist
error, ignoring other errors such as read only file system, as reported
in this bug.

In this patch, updating the SD role is back to the _domainStateChange
method, running only in the SPM host, avoiding this way to affect the
SPM start flow.
Note that since this code runs each time a domain becomes active, if
it fails to fix the domain role, it will try again on the next time the
domain becomes available.

Change-Id: Id13744a155238d63118c7de0835d54f0a5647351
Bug-Url: https://bugzilla.redhat.com/1509654
Signed-off-by: Fred Rolland <frolland@redhat.com>