virt: metadata: simplify helper functions
We added two helper functions to let the code extract the metadata
from a device object:
- get_simple_metadata
- get_nested_metadata

Those two functions have two issues:
1. the name is misleading and
2. we have two of them, forcing the programmer to distinguish
between 'simple' and 'nested' metadata with not immediately obvious
hints about where to draw the line.

In this patch, we unify the two functions and rename the resulting
one using a more intent-revealing name. We choose to always use
utils.picklecopy to copy the data and cover all the use case.

We believe the performance loss is negligible, and that the increased
convenience offsets this loss.

There is no intended change in behaviour.

Change-Id: Ia7244c38f3eb93ae7980fcd2b252415e5efd3ff4
Signed-off-by: Francesco Romani <fromani@redhat.com>