momIF: change the way we connect to MOM
Previously, we used to import MOM just to learn about the path
of the unix domain socket we need to use to talk with MOM.
To remove any remaining dependency on the MOM code,
we duplicate this information on the Vdsm configuration file.

When configuring the host, for example using ovirt-host-deploy
or ansible, the configuration agent can easily override the Vdsm
default (which matches the current MOM default), reading data
from the MOM configuration file.

Change-Id: I3b86b7cbce43d59500bba82cf00ed01f3d4aa7f5
Signed-off-by: Francesco Romani <fromani@redhat.com>