config: change default zero method to blkdiscard
This patch changes the default zeroing method from "dd" to "blkdiscard".
The new default method performs significantly better than "dd" as it may
use storage offloading if it's supported by the storage array.

To show the improvement in the performance, I ran the following test:
- Created 2 block domains - one on xtremio and the other on netapp.
- Created 2 VMs (each one had its disks on a different domain):
  * Created 5 disks of 50G (with Wipe After Delete enabled) for each VM.
  * Took 9 snapshots of each VM.
  Eventually, each VM had 5 LVs of 50G + 45 LVs of 1G, resulting in 50
  LVs of 295G in total.
- Removed each VM separately and measured how many seconds it took vdsm
  to zero all the volumes, using dd with a step of 1M and blkdiscard
  with a step of 2M.

The results are as follows:

Zeroing Method   xtremio   netapp
---------------------------------
blkdiscard(2M)   33        488
dd(1M)           2719      2745

That is, 82% improvement in the performance when zeroing disks on
netapp, and 98.8% improvement when zeroing disks on xtremio.

Also, there was a significant improvement in the network usage with when
using blkdiscard, ~0%-15%, as opposed to ~ 94% when using dd.

Here's a graph that shows the results:
https://imgur.com/a/irspB

Change-Id: I579b0ccc37cd4781272e8284d8948da9606ea630
Bug-Url: https://bugzilla.redhat.com/1475780
Bug-Url: https://bugzilla.redhat.com/872530
Signed-off-by: Idan Shaby <ishaby@redhat.com>