Support memory images on different storage domains
Up until 4.2 the hibernation volumes always resided on the same storage
domain and were represnted by a string of comma separated UUIDs.
From 4.2 onwards having the hibernation volumes on different storage
domains is supported and due to that we have to intrduce new parameters
for VM.create that will allow it and replace the old string.
This patch adds that capability while leaving support for the old string
for backwards compatibility purposes.
The tests were also changed as the existing tests didn't really check the
actual parsing but only whether the hibernation string was popped from the
parameters

Bug-Url: https://bugzilla.redhat.com/1150245
Change-Id: I05ae46ebf27fb075ca90fcd8c6da059ee278ca10
Signed-off-by: Tal Nisan <tnisan@redhat.com>