api: Log VM id on API.VM calls
@api.logged decorator helps to log API calls consistently.  However it works on
API objects rather than with actual API call arguments.  That may lead to loss
of important information if some arguments are removed before reaching API
object methods.  It happens with VM id, which is stored in the corresponding
API object but no longer present in arguments of its methods.  That makes
tracking VMs in the log difficult.

This patch enhances @api.logged with the possibility to extract and log
additional information from API objects and uses the enhancement to log VM ids
in API.VM calls.

Change-Id: I9b58eeb915858b031f38d8a85e3fd185896d5899
Backport-To: 4.2
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>