gluster: Re-enable tests depending on blivet Fedora
Importing blivet has been broken for a while on Fedora, but the
recent version available in Fedora 27 seems to fix it, for python 2 at
least.

This patch requires that version re-enable the tests that depend on it
for python 2.
The python 3 implementation still seems to break the current code
(e.g., LVMThinLogicalVolumeDevice that is required by
lib/vdsm/gluster/storagedev.py does not exist in the python 3
implementation delivered by python3-blivet-2.1.11-3 currently
available in Fedora 27).

Change-Id: Ie9d147d034b15218f4146afc71979b229897c901
Bug-Url: https://bugzilla.redhat.com/1530724
Signed-off-by: Allon Mureinik <amureini@redhat.com>