vmdevices: refresh only disk devices
The lease refresh conflicts with the work we do in fixLeases.
If we use both mechanisms:
- fixLeases will fill the placeholders with the right data
- the refresh code will replace the fixed leases XML
  with incorrect data rebuilt from the original XML, before
  fixLease could have fixed it.

To fix this issue and the possible conflicts with graphics
and network devices, we make Vdsm by refresh only disk
devices, avoiding any overlap between refresh and placeholders.

We also kill the devel tunable which turned out just adding confusion.

Change-Id: I87116fdf73dc3bd33b7e964365d53aaa76b14607
Bug-Url: https://bugzilla.redhat.com/1547479
Backport-To: 4.2
Branch-Name: fix-drive-leases
Signed-off-by: Francesco Romani <fromani@redhat.com>