spec: lower the libvirt requirements
On Fedora 27, the newest available libvirt release is 3.7.0.
That means that the virt-preview yum repository is required
to satisfy the libvirt requirement of 3.10.0-x.

However, virt-preview is currently only available for the
x86_64 architectures preventing the update of vdsm
on ppc and s390 systems.

This change lowers the requirement for all architectures to
the libvirt level available if built for Fedora 27 which
already contains the fix needed by vdsm.

Change-Id: Ib531599088a6ab61d2c02ca3483f85bb1328aec9
Signed-off-by: Viktor Mihajlovski <mihajlov@linux.vnet.ibm.com>