virt: Kill killable VMs paused for too long
We already kill VMs with ResumeBehavior == KILL if they are going to be resumed
after being paused for too long.  However we perform that action only on
resume.  If a paused HA VM is not resumed, it keeps being paused.

This is not a good behavior for HA VMs, since they do not run while being
prevented from starting on another host.  So we add a periodic check that kills
VMs with ResumeBehavior == KILL paused on I/O error for too long.

Note that we kill paused VMs with both EIO and EOTHER pause codes.  EIO is set
on I/O error.  EOTHER is set on recovery in Vm._readPauseCode and, according to
the comments there, there may be some some reasons for that.

Change-Id: Iac28225302b5d3ddd890ddaca0eff5e9e7706c8c
Bug-Url: https://bugzilla.redhat.com/1540548
Backport-To: 4.2
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>