virt: domxml_preprocess: make devices if missing
We want to call the placeholders handling function very early in the
Vm initialization flow, when the device objects are not yet available
- not they should be, because this will defeat the purpose of having
the placeholders code.

So, we add one helper function to build the devices out of the XML,
which will be used also in the dehibernation flow when we will move
that code in this module.

Change-Id: I565fde0e1b36a6ed47677292b6573d8f664e619e
Related-To: https://bugzilla.redhat.com/1547479
Backport-To: 4.2
Branch-Name: domxml-preprocess
Signed-off-by: Francesco Romani <fromani@redhat.com>