nfs: Disable NFSv3 locks
Since 2.10, qemu takes file locks on images. This breaks HA VM failover
when using NFSv3, and may prevent starting a VM on another host after
the original host was powered off and the user confirmed the operation.

We use now the 'nolock' mount option for NFSv3 or Auto-Negoticate
mounts. This disables NFSv3 cluster locks. The locks taken by qemu will
have effect only on the same host when the VM is running. This provides
the same protection that we have with block storage, when qemu locks the
logical volume. The 'nolock' option does not effect NFSv4 locks.

If users need NFSv3 remote locks and do not care about the issues, they
can override this per storage domain using the 'lock' mount option in
engine.

4.1 notes:
- Remove the "id" argument when creating MountConnection.  This argument
  was added in 4.2.

Change-Id: I25e6184bda53c2566d175681d6db1c24bcf88444
Bug-Url: https://bugzilla.redhat.com/1550127
Signed-off-by: Nir Soffer <nsoffer@redhat.com>