virt: Handle libvirt error in Vm._readPauseCode on recovery
_vmDependentInit exceptions are swallowed on recovery (we can't ignore or
discard VMs already present), which may leave a VM in inconsistent state.
Let's ignore _readPauseCode libvirt exceptions on recovery to prevent the
problem.  Ignoring a contingent _readPauseCode libvirt exception may mean
e.g. that a VM paused on I/O error won't be resumed once the storage error
disappears, but that's just one of the problems already present when the
exception is not ignored.

We still forward the exception in non-recovery flows, in which it is propagated
and should be handled as VM start failure properly.

Change-Id: I0c9cd2f4a7e8f7f4a6f2a305451a7a7ab15ceb4a
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>