lib: executor: print duration as float
The duration of one executor.Task can actually be a floating
point number, and we should represent it as such.
This make the outcome of repr(executor.Task) clearer, because
we already represent the timeout as floating point number.

Please note that we use '%s' for the timeout because it
may also be None.

Change-Id: I0637000f95eba2113c457e9f35fb6977b60f3b2a
Topic: janitorial
Backport-To: 4.2
Signed-off-by: Francesco Romani <fromani@redhat.com>