vmstats: initialize cpuUsage
We compute cpuUsage out of cpu.user and cpu.system,
but we don't inialize this value if no data is available.

The lack of this value may confuse MOM.
In most cases, we will update the value with real data,
so we will not lie to Engine anyway.

Change-Id: I87500a8b8581394e6b0c4d882999b91ddbcc9785
Backport-To: 4.2
Branch-Name: janitorial
Signed-off-by: Francesco Romani <fromani@redhat.com>