spec: Require nfs-utils to support minorversion
Platform bug 1547681 describes an issue in nfs-utils where the
minorversion mount option is not honored in nfs mounts.

This patch is a quick fix to require an nfs-utils version that fixes
that bug for EL 7.4.z.
A more comprehensive fix could be to drop the parsing of the version
the engine send and just pass it verbatim done to the mount command.
If that is done, this patch can be reverted.

Change-Id: I17df1bc4fba285809c554f0ca23716c229bc08e2
Bug-Url: https://bugzilla.redhat.com/1522692
Related-To: https://bugzilla.redhat.com/1547681
Signed-off-by: Allon Mureinik <amureini@redhat.com>