vm: don't replace device XML, only hooks output
The device hooks may alter the device XML snippets.
This is part of the contract between Vdsm and hooks, so there
is little we can do with that: the XML Engine (or any any client)
provided can be altered by hooks.

Nevertheless, Vdsm should strive to minimize those changes.
This is the first patch in a series that wants to achieve that,
in order to get as close as possible to Engine XML passthrough.

This patch removes the generic device XML replacement code, and
updates it instead only for disk devices, or if one hook produced
some updated XML.

Change-Id: I2536338b931a7f780a969cb2e14f132c88b61937
Topic: domxml_update
Signed-off-by: Francesco Romani <fromani@redhat.com>