net: ovs: Support setting the mtu on an ovs base net
Setting the MTU on an OVS base network has not been supported so far.
With this patch, VDSM networking supports setting the mtu with some
limitation in mind.

As the OVS host networking bridges are aimed be used only for
supporting the physnet of an OVN network, and vnics are not planned to
be attached to it directly, the mtu settings are limited to the bridge
southbound and northbound ports, where the sb is to use the maximum mtu
of all the nb defined on it.

When reporting a network mtu, it is gathered from the nb ports mtu.

https://bugzilla.redhat.com/1598461
Change-Id: I83d5679635aa94ed706ea67cffb01cd2184d3660
Signed-off-by: Edward Haas <edwardh@redhat.com>