net: On IPv6 address add, do not flush existing addresses
Adding an IPv6 address using iproute2, caused the existing IPv6
addresses to be flushed first and only then new config was added.
This approach has been found problematic recently, as the flush did not
finished its async operation, and tests failed on adding the address
(claiming that an identical one already exists).

The flush step at the ip.address level is too limiting and should
have been handled at the callers.
Attempting to add an IP address and if already exists to log it and
continue seems more appropiate.

Change-Id: I2274350e988daeaee8dff5a9bc2f8286ce87ae7f
Signed-off-by: Edward Haas <edwardh@redhat.com>