virt: Don’t look for unknown VMs indefinitely
clientIF tracks VM ids seen in libvirt lifecycle event handler and not
belonging to any known VM.  A periodic jobs asks libvirt about those ids and if
they are known, it creates corresponding external VM objects.

If an id doesn't belong to a VM known to libvirt, the periodic job keeps asking
libvirt about it and logs related debug messages.  This is unnecessary – if the
id is unknown to libvirt, it can't belong to an external VM and it doesn't make
sense to ask about it anymore.

Actually the periodic task should handle each unknown VM id, either by creating
a corresponding external VM or by ignoring the id (for various reasons).  The
only exception is when it encounters an unexpected libvirt error, then it may
make sense to process the VM id again.

Note that we should be careful about not losing an unprocessed unknown VM id,
from both lookup_external_vms and event dispatcher running at the same time.
This patch handles that without wrapping libvirt calls by
clientIF.vmContainerLock.

Change-Id: I1d9ac17de545823a6dab4e87b0018da3aaa40ea7
Bug-Url: https://bugzilla.redhat.com/1552713
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>