virt: storage: minimal changes to the drive XML
Because of the per-host changes that Vdsm needs to do
on drive configuration, it used to replace entirely the
XML snippets for drive configuration.

This is awkward, and forces developers to update both Engine
and Vdsm to support new features (or anyway change the drive
configuration), often leading to surprising behaviour.

This patch wants to improve this state doing minimal
changes to the drive xml, just replacing the few attributes
that we know (and we allow) to change per-host basis:

- drive path
- drive type (file vs block...)
- drive format (raw vs cow)

this includes the attributes that can change because of
the after_disk_prepare hook point - aka the localdisk hook.

Change-Id: I3fd117c7eddce3e0e075766adeaceaf165364d60
Backport-To: 4.2
Topic: domxml_update
Signed-off-by: Francesco Romani <fromani@redhat.com>