tests: Fix flaky time based test
Since we merged the new threaded-logging tests, we have a new random
test failure in the CI:

Traceback (most recent call last):
  File "/home/jenkins/workspace/vdsm_master_check-patch-fc27-x86_64/vdsm/tests/common/logutils_test.py", line 313, in test_slow_handler_sync
    self.assertGreater(max(workers_time), 0.9)
AssertionError: 0.8729510307312012 not greater than 0.9

I think the reason for this failure is interrupted time.sleep(). This
function is documented to return early if the syscall was interrupted by
signal. Looking at python timemodule.c, no exception is raised in this
case.

The sleep code was changed to sleep at least delay seconds in case of
silent interruption.

To make it easier to debug, we print now the worker threads times. The
print is visible only if using the -s flag, or test fails.

Change-Id: Ied84ee06935391bd613db47db0e70fb696b52396
Bug-Url: https://bugzilla.redhat.com/1516831
Signed-off-by: Nir Soffer <nsoffer@redhat.com>