virt: snap preview: amend data only for disks
When we preview a snapshot, we need to amend the drive
XML configuration with fresh data provided by Engine.

We should do that only for drives which support snapshotting,
thus non-LUN drives.

+++

The current code also attempt to amend the XML also if the
device is '' (empty string). We believe this is defensive programming,
no longer useful nowadays. Thus, we remove this as well.

Change-Id: I9c76472da9d89f4064bc8d875aa63e01923ec367
Topic: change-disk-fixes
Backport-To: 4.2
Bug-Url: https://bugzilla.redhat.com/1566948
Signed-off-by: Francesco Romani <fromani@redhat.com>