virt: snap preview: look by serial, then for alias
When we preview a snapshot, Engine sends update drive information,
that Vdsm must use instead of the stored domain XML.
We need to find the drive to update, and amend its information
(or, in a not so distant future, to replace it).

To identify the drive, we used the drive `serial`, which is not
available for all drive types. Incidentally, the drive which don't
support `serial`, also don't support snapshots, so no functionality
is lost.

Starting with libvirt 3.9.0 and Engine (circa) 4.2.2, we can use
client-provided device aliases, and Engine started to send them.
Then, we can use aliases to uniquely identify the drive to update,
in a way both more robust and more generic than the drive `serial`
alone.

For the sake of backward compatibility, we still try to lookup
the drive by serial first, and only if this fails we fall back
looking by user-alias.

Change-Id: If4a2245673dfb70135415795e53c2c6ce31453e7
Topic: change-disk-fixes
Backport-To: 4.2
Bug-Url: https://bugzilla.redhat.com/1566948
Signed-off-by: Francesco Romani <fromani@redhat.com>