net: When mocking builtin 'open', use 'create=True'
The test_daemon_cmdline_with_last_arg_as_hostname test is failing on
some platforms (detected on RHEL 7.4), claiming that the 'open' function
does not exist when attempting to mock it.
It is not clear why on CI or other platforms it passes without this
error, it just appeared on a new development env. (python, pytest and
python-mock used the same versions)

Adding to the mock patching the option to create the target even if it
is 'missing'.

AttributeError: <module 'vdsm.network.ip.dhclient' from
'/home/svetah/workdir/vdsm/lib/vdsm/network/ip/dhclient.py'> does not
have the attribute 'open'

Change-Id: Ibabadfa55bdfd174dbbb4ec06466c0d142881812
Signed-off-by: Edward Haas <edwardh@redhat.com>