net tests: Drop inheritance from all functional test
Functional tests have been inheriting the helper class NetFuncTestCase for
accessing the API and asserting.
This patch is replacing the inheritance with a module level object.

It is possible now to create a constructor and allow in following
patches to add a flag, that control with what API the tests should run.

The class name was changed to NetFuncTestAdapter.

Change-Id: I1681591dfd6e80c0331c24a710e87137b3a3de41
Signed-off-by: Sveta Haas <sveta.haas@gmail.com>