net tests: Adjust package fixtures and duplicate to integration
With the move of the integration tests to their own package, their
fixtures should also be moved. The existing two fixtures are handling
the bond options mapping and the cleanup of ip rules in case any are
left behind.
Currently both old and new fixtures are needed as the nose tests still
call the bond options. When all tests will move, the nose fixtures
will be droped.

With this patch, the bond options mapping is changed a bit for the
tests, triggering it once per session run and forcing bond options
mapping to be executed whenever possible (sysfs write access).

It has been found that the tests may run on different slaves which have
different kernel versions, therefore different bond options defaults.

Change-Id: I02d61b6bef5dfc4b2de54c03b0f7ffd3d3eee978
Signed-off-by: Edward Haas <edwardh@redhat.com>