domxml_preprocess: cdrom attr may be missing
When we restore a hibernated vm, the attribute of the <source>
element which identifies the image path may be missing.
For example, using ISO images on NFS domain, this may be a valid
source element returned by libvirt:

  <source startupPolicy="optional">

note the lack of "file" attribute (being the image on NFS), compared
with what Vdsm builds:

  <source file="" startupPolicy="optional">

This patch makes sure the device replacement code can deal with missing
attributes, and adds the corresponding test.

Change-Id: Ia4dac75678b58b2f11f3cced2a88a78b17a76488
Backport-To: 4.2
Topic: domxml_update_fixes
Bug-Url: https://bugzilla.redhat.com/1561010
Signed-off-by: Francesco Romani <fromani@redhat.com>