net tests: Add a timeout option to the ovs driver
When executing an OVS command, allow the caller to specify a timeout
argument, which limits the time the command can block.

In some cases, the connection to the ovsdb service may hang and with it
hang the whole VDSM. Specifying a timeout limit will allow VDSM to fail
faster.
The hang state has been detected during the integration tests on the OVS
driver, no similar problem has not been detected in production.

In this patch, by default no timeout exists. Only the OVS setup/teardown
fixtures make use of it for the time being.

Change-Id: I2b7f2f680fd5bd58b16107e80add08ef8299f1a5
Signed-off-by: Edward Haas <edwardh@redhat.com>