net tests: Extract bridges cleanup from the OVS fixture
The OVS fixture helper includes in its teardown the removal of all bridges
used in the tests. In order to optimize the fixture, the cleanup step is
extracted out to its own function, which needs to be called explicitly
on each test teardown (done through the class fixture).

This change is introduced in order to allow in the next patch to convert
the OVS service fixture to a module scope.

Change-Id: Id0c44e6073dbaa6fa936f8a2bfedbef2bfe6564e
Signed-off-by: Edward Haas <edwardh@redhat.com>