net tests: Use OVS service as module scope fixture
The OVS service fixture is changed from function scope to module scope.

The frequent start/stop of the OVS service (on a per function basis) is
suspected to cause issues of hanging the ovs-vsctl command.
If it is the cause or not, it does not make sense to do so per each
test.

Change-Id: Ifb8edc81dfc78c20caeab8b46cf4509848ef0dce
Signed-off-by: Edward Haas <edwardh@redhat.com>