net tests: Migrate OVS unit tests to pytest
An unit package has been created in the network package.
The OVS unit tests have been moved into unit.ovs package.

Note: With this change assertNotRaises has been dropped,
letting the tests fail (on error) if an exception is raised.
assertNotRaises is a home made version which does not exist under
nose or pytest.
It had to be dropped or recreated due to the preference of not
depending on non-network helpers.

Change-Id: Ia5a02a7b852ad41f9417a871e07e47f460538fcd
Signed-off-by: Edward Haas <edwardh@redhat.com>