net tests: Handle in-flight device change in test
With the recent changes in the integration tests, the order and timing
of the test run changed and with it a race in the NM driver usage has
surfaced.
The test now ignores errors when attempting to access a device property
which is no longer accessible (due to changes in the device state).
The test requires only that at least one device is detected and fully
processed.

Change-Id: Ia9634efe98c27381efeb7e7ab71430d9f72c14f3
Signed-off-by: Edward Haas <edwardh@redhat.com>