net: Unify legacy and ovs setup validations
Existing OVS setup validations includes generic VDSM networking
validations which fits both legacy and ovs.
The same can be said about most of the legacy setup validations.
Therefore, in this patch both are unified, keeping only the DPDK
interface validation for the legacy setup.

In order to help with the review process, in this patch the unified
outcome is moved into the ovs.validator.
In the following patch, the ovs.validator will be moved to the
netswitch.validator which already has some common validators.

Change-Id: I20cd66044affe8e03af1673e57a4e46d72c2bcbe
Signed-off-by: Edward Haas <edwardh@redhat.com>