virt: dom_desc: add get_device_elements_with_attrs
A common usage patter in virt code is to fetch all the device
elements which have some attributes set to some other values.

This patch enhance DomainDescriptor to support this
operation. The new method is a superset of get_device_elements,
which will be phased out in a future patch.

Change-Id: I61b8076a2e4705e4c6deb74c5df94082d9b7fd02
Backport-To: 4.2
Topic: domxml_update_preview_snap
Signed-off-by: Francesco Romani <fromani@redhat.com>