Increase memlock limit on POWER
virt-sparsify called from Vdsm may fail on POWER with errors like

  libvirt:  error : cannot limit locked memory to 18874368: Operation not permitted

Apparently the default memlock limit is not applied to virt-sparsify
invocation, so we must set it explicitly in systemd configuration.  We use the
same limit as in /etc/security/limits.d/95-kvm-memlock.conf, i.e. 64 MB.

Since the problem is related to POWER hash page table, it's specific to POWER
and we set the limit in systemd only for POWER.

Change-Id: I2e920f0e3dae27a45a934fdfeca784961deec22a
Bug-Url: https://bugzilla.redhat.com/1447300
Backport-To: 4.2
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>