net tests: Create test adapter through a module fixture
Each module has been creating its own test adapter at the module
global level. Unfortunately, doing so seems to be problematic when
attempting to run the tests a second time (and forward).
From an unknown reason, pytest somehow creates python
cache files (*.pyc) which do not allow the proxy connection
to be re-established.

This patch is suggesting to resolve this by a workaround that seems to
work: initialize the module global adapter using a module level fixture
which is automatically triggered (autouse).

Change-Id: I54957c38aba1e1ff350cbd0babdccebe576f093d
Signed-off-by: Edward Haas <edwardh@redhat.com>