virt: vmxml: lib: add fromstring() helper
In python2, ElementTree doesn't work with unicode data:
https://bugs.python.org/issue11033

This became evident attempting to run the virt testsuite
with pytest. I'm not sure why it not surfaced before.

In this patch, we add a fromstring() method in the xmlutils
module, which takes care of transparently encoding the
data before to call ElementTree.
Since vmxml.parse_xml() is now redundant, this patch also removes it.

Change-Id: Ifc275a6bda7910a6970d6cf56b1ef2007e444559
Topic: pytest-virt
Signed-off-by: Francesco Romani <fromani@redhat.com>