net: Refactor SRIOV setup flow and its persistency
SRIOV setup flow has been refactored in order to merge its persistence
with the other network and bond entities.

In the existing setup flow, the SRIOV numvfs values have been persisted
in a parallel flow, externally to the netconfpersistence module, causing
the startup/safe config to be lost on boot.

The proposed changes include the migration/upgrade of the sriov
configuration which is persisted to a new format under the
netconfpersistence module.
The netupgrade module is responsible to upgrade the configuration to the
new scheme.

Change-Id: Ib4390d6d27c2f49ccc847ae155798e52bea099ce
Bug-Url: https://bugzilla.redhat.com/1563165
Signed-off-by: Edward Haas <edwardh@redhat.com>