net: Adding to the OVS driver a default 5sec timeout
In case the ovsdb is not responding for more than 5sec, we conclude that
something very bad happened and should abort.

In recent tests run with OVS 2.9, we have seen this happening from time
to time, causing the test run to get stuck until the CI timeout.
We would prefer to fail with an exception.

Change-Id: I27fd6128375550841c316ad3b18a802b73cf1b67
Signed-off-by: Edward Haas <edwardh@redhat.com>