build: Made vdo-gluster arch-dependent
TL;DR As vdsm-gluster package now requires vdo on x86 architecture,
it can't be arch independent anymore.

We encountered a problem with brew and this patch [1]
brew is not happy with noarch package, acting differently
on different architectures. So we need to make gluster-vdsm
arch dependent.

As a temporary solution we make vdsm-gluster arch dependent,
to not to delay build for another one week. After that
we will make it arch-independent again and create a
virtual vdsm-gluster-vdo package, that will provide
vdo dependency on x86 arch.

[1] https://gerrit.ovirt.org/#/c/88884/

Change-Id: I7648af1547206c7ef68a0ce5c5ebf0fb37cc50ae
Signed-off-by: Denis Chaplygin <dchaplyg@redhat.com>