tests: rewrite py-watch and make it easier to use
Before this patch, py-watch was watching a pre-existing pid. This forced
the caller to spawn the watched process in the background, which was
cumbersome to use.

With this patch, py-watch spawns the watched process according to the
command-line arguments it receives:

  py-watch $timeout command args

is simpler (i.e. from tox.ini) than

  command args & pid=$! ; py-watch $pid $timeout

The current implementation spawns the watched process in a fresh
session, so that py-watch can kill it together with any process it might
have forked.

py-watch is rewritten in Python which is more of a programming language,
comparing to bash. Proper tests are written to document py-watch usage
and prove its effectiveness.

Change-Id: Ia47c878aa562ffd2e96d46efec43e9fdc28a138c
Signed-off-by: Dan Kenigsberg <danken@redhat.com>