virt: use 'qemu' domain in the placeholder XML
Early in the creation flow, Vdsm prepares a minimal, placeholder
domain to be used only to allow to save the domain metadata before the
domain is really started (virDomainCreateXML).
Shortly after, the VM will be re-defined using the real, meaningful
domain, and the placeholder will never seen again.

Up until now, we used <domain type='kvm'> for the placeholder domain,
because it is most similar to the real XML we will create, and we wanted
to avoid any unneeded difference.

Turns out that type='kvm' breaks the faqemu hook; thus we
switch to type='qemu', only for the placeholder XML.

Change-Id: Ifa5cac9810fca044de6a447b6abda01c47122ff6
Topic: master
Signed-off-by: Francesco Romani <fromani@redhat.com>