net: Drop the ovs netinfo 'nics' and 'bond' keys
In the OVS caps(netinfo) report, the network section included new keys
which did not exist with the original linux bridge switch report.

As we aim to unify the two reports such that there will be no need to
handle the netinfo differently between switch types, the 'nics' and
'bond' keys are dropped in favor of 'southbound'.
The southbound interface is the one that connects the network to the
host externals.
It replaces 'bond' by requiring the user of the report to check if the
southbound is in the bondings section and replaces nics by handling the
bond slaves though the bondings information and not as pert of the
network one.

In the following patch, the southbound key will be added to the linux
legacy report as well, allowing a unify handling of the netinfo/caps.

Change-Id: I43d06a858e42e52c3940832107fae8dc5db71eb1
Signed-off-by: Edward Haas <edwardh@redhat.com>