net: Add 'southbound' to the network legacy switch type report
The 'southbound' key/value was added to the network netinfo of OVS based
networks, in this patch it is added to the linux bridge switch type
networks report as well.

The functional tests adapter has been updated, to assert on this
parameter as well.
Note that the previous assertSouthboundIface assertion was named
incorrectly, it should have been named assertNorthboundIface. This has
been fixed in this patch.

Change-Id: I294ebd19d4030ff5f4dcf7df3b1384ade770a208
Signed-off-by: Edward Haas <edwardh@redhat.com>