virt: domxml_preproc: CDRoms may lack driver attrs
It seems that libvirt >= 3.9.0 may return a valid XML semantically
equivalent to what Vdsm submitted, but lacking fields that used
to be there.
We already seen another instance of this in change
Ia4dac75678b58b2f11f3cced2a88a78b17a76488

This patch makes Vdsm deal with the fact the fields it looks for
may be absent.

To make sure we don't hit more bugs like this, we will need to
run OST on EL7.5, libvirt >= 3.9.0

Change-Id: I360a5ee9d73f8bfb4ef5a2f4a97051fd8edf4c4e
Topic: domxml_update_fixes
Bug-Url: https://bugzilla.redhat.com/1567617
Backport-To: 4.2
Signed-off-by: Francesco Romani <fromani@redhat.com>