automation: check-patch: slave and in-mock log collection
Suring check-patch, multiple integration tests are running which utilize
the slave services (like NetworkManager and openvswitch).

Recently, with the new OVS 2.9 version, the tests have become unstable.
In order to assist in the debugging of the various services, a log
collector is added, collecting the logs from both the slave and the
local mock logs.

Change-Id: I24ea67c2ce77a881fca1c6ab10919dbeaeb6bf99
Bug-Url: https://bugzilla.redhat.com/1568268
Signed-off-by: Edward Haas <edwardh@redhat.com>