virt: domxml_preproc: minimal update for floppies
To properly support sysprep floppies[1], we need to amend the path
of the floppy disk elements. But we need to update just the path,
because other attributes either don't matter for floppies, or may
just be absent, because the folppy configuration is expected to be
different - and simpler - with respect to cdroms and disks.

+++

[1] sysprep is the only last meaningful use case for floppies at present
day.

Change-Id: If6e00f2eb5042ee6d4a27b5e73adb41c0b802fd2
Topic: domxml_update_fixes
Bug-Url: https://bugzilla.redhat.com/1567617
Backport-To: 4.2
Signed-off-by: Francesco Romani <fromani@redhat.com>