check-patch: fail a patch if it lowers network pytest coverage
Our current test coverage is only 42.18% in the network subpackage.
It is quite lame, but it should not get any lower.

If you lower networking coverage, expect a failure such as

    FAIL Required test coverage of 42% not reached. Total coverage: 26.52%

This patch makes it a bit harder to lower our coverage by mistake.
However, there may be good reasons to allow lowering the coverage, for
example, if you have an ultra-urgent bug to fix in a not-yet-covered
module, or if you move tests to pytest, which currently not accounted
for.

If someone adds code, has a very good reason not to add tests for it,
and manages to convince one's maintainer, the maintainer should
explicitly approve the reduction of 42% to something lower.

Change-Id: I8ea2590a875314783cda51b64a5bb0683b2d4621
Signed-off-by: Dan Kenigsberg <danken@redhat.com>