virt: properly store Drive specParams
In the vm.conf compatibility flow (Engine <= 4.1),
Vdsm must add the data it needs to the metadata section.
We use the metadata to store the informations that Vdsm needs,
but which are  not represented in the domain XML.

In the newer XML-based flow (Engine >= 4.2), that information
is filled by Engine and never changed by Vdsm.

We store three informations in the Drive.specParams:
1. payload device data
2. iotune settings - but we restore them from the domain XML
3. iothread pinning - but we restore this from the domain XML.

Due to a bug in the code, Vdsm failed to store the drive specParams
into the metadata XML. This cause a bug when we have payload devices:
migrations could fail, and recover is partial.

Saving all the three of them leads to unnecessary data floating around,
leading to possible bugs or confusion.
For this reason, this patch simplifies the metadata code to save only
the payload data, ignoring iotune info.

Change-Id: Ifad7ed0dbbca1fdb61d740573f479804bb5fedbc
Backport-To: 4.2
Topic: domxml_fixes_drive_spec_params
Signed-off-by: Francesco Romani <fromani@redhat.com>