vm: Do not remove replicate attribute if abort failed
If self._dom.blockJobAbort(drive.name, blockJobFlags)
raises an exception, we still remove the 'diskReplicate'
in the finally clause.

In newer engines (4.2+), when we try to abort the replication
it will throw a ReplicationNotInProgress exception which will
make the engine try and cleanup the target disk (a failed attempt).

Instead, it should reach the self._dom.blockJobAbort with
with the 0 flag which aborts the blockjob, and then perform
the cleanup.

In older versions it will simply fail without trying to
perform a cleanup.

Change-Id: I96af7958fc87719381511261d95120fd7eb50eb1
Bug-Url: https://bugzilla.redhat.com/1502083
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>