virt: sampling: get stats only for running VMs
Currently we ask libvirt for the stats about all the VMs it
knows about.

With this patch, we query only running VMs.

There is little to none reason to do otherwise, because Vdsm clients
don't care about stats for non-running VM, which are meaningless - at
best.

Previously, all VMs were transient, so for all intents and purposes,
only running VMs existed on a given oVirt host.

Please note this patch is meant to improve, but not solve, the
rhbz#1537148.

Change-Id: I14000576b409a88740147e192d1f883e2021c68d
Topic: memstats-running
Backport-To: 4.2
Bug-Url: https://bugzilla.redhat.com/1537148
Signed-off-by: Francesco Romani <fromani@redhat.com>