net: Fix sriov configuration upgrade
With the introduction of the new sriov persistency, the upgrade process
handled both the running and persistent config in the same way, trying
to upgrade the old sriov config from the persistent folder (only).

With this change, the old configuration is taken from both the persisted
and running (staging) folders.

Change-Id: Ie37d6b73a3d561ecbe91630fb34c8acbf65bde93
Bug-Url: https://bugzilla.redhat.com/1563165
Signed-off-by: Edward Haas <edwardh@redhat.com>
(cherry picked from commit 038b55813ed67a419aa2763637bcc345564900c4)