clusterlock: Fix lease status check
This patch comes to address an issue where a resource lease is
considered owned, despite the fact the host which acquired it had
disconnected from the storage and vdsm is restarted by sanlock.
In this case, the resource lease is no longer "owned" since it was
acquired by a different pid. So currently:

* If a host was disconnected from the storage and reconnected,
the lease will be considered owned forever
* If a host was disconnected from the storage and did not reconnect,
the lease will be considered owned forever

The issues that may rise from this scenario are jobs that are
"forever" considered running by the engine because the related lease
is considered owned even after the connection to the storage is
restored.

This patch addresses the issue by comparing the lease
generation and host generation and checking the owning host's status.
The lease will be reported as owned only if the lease generation and
host generation or the host's status is either "live", "fail" or
"unknown". In any other case the lease will be considered free.

A new exception was introduced to handle failures when we attempt to
acquire but acquireHostId was not called:
TemporaryFailure: Cannot inquire Lease(name='SDM', path=u'/path/to/leases', offset=1048576): (11, 'Sanlock get hosts failure', 'Resource temporarily unavailable')
2018-06-14 07:44:54,767-0400 INFO  (jsonrpc/7) [storage.TaskManager.Task] (Task='uuid') aborting: Task is aborted: u"Cannot inquire Lease(name='SDM', path=u'/path/to/leases', offset=1048576): (11, 'Sanlock get hosts failure', 'Resource temporarily unavailable')" - code 100 (task:1181)
2018-06-14 07:44:54,767-0400 ERROR (jsonrpc/7) [storage.Dispatcher] FINISH getSpmStatus error=Cannot inquire Lease(name='SDM', path=u'/path/to/leases', offset=1048576): (11, 'Sanlock get hosts failure', 'Resource temporarily unavailable') (dispatcher:84)
Traceback (most recent call last):
  File "/usr/lib/python2.7/site-packages/vdsm/storage/dispatcher.py", line 71, in wrapper
    result = ctask.prepare(func, *args, **kwargs)
  File "/usr/lib/python2.7/site-packages/vdsm/storage/task.py", line 108, in wrapper
    return m(self, *a, **kw)
  File "/usr/lib/python2.7/site-packages/vdsm/storage/task.py", line 1189, in prepare
    raise self.error

Change-Id: I27eb41d466f6478898fc8ece7333d718a94d9957
Bug-Url: https://bugzilla.redhat.com/1565040
Bug-Url: https://bugzilla.redhat.com/1553985
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>